ml-server/pyproject.toml
[project]
name = "ml-server"
version = "0.1.0"
description = ""
authors = [
    {name = "Praveen Kumar",email = "praveenkumar.e2022@vitstudent.ac.in"}
]
readme = "README.md"
requires-python = "3.11.11"
dependencies = [
    "fastapi (>=0.115.8,<0.116.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
    "onnx (>=1.17.0,<2.0.0)",
    "onnxruntime (>=1.20.1,<2.0.0)",
    "numpy (>=2.2.3,<3.0.0)",
    "huggingface-hub (>=0.29.0,<0.30.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "transformers (>=4.49.0,<5.0.0)",
    "pillow (>=11.1.0,<12.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "pandas (>=2.2.3,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
